Use bootstrap-sass for our Bootstrap dependencies
Since we'll be using RailsAdmin, which includes a dependency on the
bootstrap-sass gem, we should be using bootstrap-sass in the main
app as well. This gives us better compatibility with the asset
pipeline and RailsAdmin, and also helps get the mix-and-match
Bootstrap versions we currently have in master.

This is my first attempt, it was mostly a global search-and-replace
and then manually fixing anything obvious that I broke. It works for
me in Firefox 7 on Ubuntu and in Opera Mini and the native browser on
my Blackberry. It should probably be tested on other browsers and
devices. :)

This commit also moves stylesheets from app/assets/stylesheets/vendor
to vendor/assets/stylesheets and refactors a few things into SASS
partials.